lpc176x: allow 16 bit SPI length requests

+++ src/lpc176x/gpio.h
@@ -28,6 +28,6 @@ struct spi_config {
 struct spi_config spi_setup(uint32_t bus, uint8_t mode, uint32_t rate);
 void spi_prepare(struct spi_config config);
 void spi_transfer(struct spi_config config, uint8_t receive_data
-                  , uint8_t len, uint8_t *data);
+                  , uint16_t len, uint8_t *data);
 
 #endif // gpio.h
